backend/requirements.txt
# ---------------------------
# Web framework
# ---------------------------
fastapi==0.117.1
uvicorn[standard]==0.30.6

# ---------------------------
# Core utilities
# ---------------------------
python-dotenv==1.0.1
typing-extensions==4.12.2
pydantic==2.8.2

# ---------------------------
# OCR / PDF
# ---------------------------
pymupdf==1.24.9
pytesseract==0.3.13
pillow==10.4.0

# ---------------------------
# ML / Embeddings
# ---------------------------
torch==2.2.2
sentence-transformers==3.1.0

# ---------------------------
# LLM orchestration
# ---------------------------
crewai==0.95.0
litellm==1.52.3

# ---------------------------
# LangChain (Gemini)
# ---------------------------
langchain-core==0.2.43
langchain-google-genai==1.0.10

# ---------------------------
# Vector DB
# ---------------------------
pinecone==7.3.0

# ---------------------------
# Observability
# ---------------------------
langfuse==2.47.0
openlit==1.30.2

# ---------------------------
# Toon format (CORRECT)
# ---------------------------
toon-format==0.9.0b1
